# Speed up factorisation for 128-bit integers