--- a/Cargo.toml
+++ b/Cargo.toml
@@ -305,6 +305,7 @@ libc = "0.2.153"
 lscolors = { version = "0.20.0", default-features = false, features = [
   "gnu_legacy",
 ] }
+machine-factor="0.2.0"
 memchr = "2.7.2"
 memmap2 = "0.9.4"
 nix = { version = "0.29", default-features = false }