--- a/src/uu/factor/Cargo.toml
+++ b/src/uu/factor/Cargo.toml
@@ -23,6 +23,7 @@ num-traits = { workspace = true }
 rand = { workspace = true }
 smallvec = { workspace = true }
 uucore = { workspace = true }
+machine-factor = {workspace = true}
 num-bigint = { workspace = true }
 num-prime = { workspace = true }