@sylvestre 2025-03-07
did you run some benchmarks ?
we like hyperfine for this ;)