@JASory 2025-03-08
I edited it to use a 64-bit which will be slightly faster for the interval, but it still probably won't be faster than GNU factor for the interval checked (0;10^7). 

The reason is trial division is faster than pollard rho for small values, but throwing more trial division at most 128-bit composites is actually a complete waste so I don't bother to optimise for that.
 
Note that half of all 128-bit composites are greater than 2^127. So a benchmark for the average input value would be much more accurate to test. 

The benchmark I ran was random input of 128-bit integers generated by openssl, which unfortunately is quite a bit harder to test when factor doesn't support hexadecimal input. (However, my algorithmically [identical fork](https://github.com/JASory/factor) does. which is what I benchmarked).