@JASory 2025-03-08
Factorisation of many numbers beyond 2^128, like  2^128+1 or 2^128+25, 2^128+35 fails. But this is an issue from the num-prime library, apparently it just gives up even though pollard-rho will eventually succeed.