@JASory 2025-03-24
> any update? thanks

Yes, it's slightly faster now, the code was simplified. Realistically beating GNU factor for tiny integers would require precomputing large prime_inverse tables which I am not interested in. As is, on my machine this implementation is faster than the current one in all intervals. 

I haven't bothered to make corrections for the CI since I have little free time now, and last I checked building for termux is going to fail until they update their rust toolchain. 

New benchmarks, using hyperfine

```
New version

Benchmark 1: seq 2 1000000 | target/release/factor > /tmp/a
  Time (mean ± σ):      3.957 s ±  0.031 s    [User: 2.819 s, System: 1.152 s]
  Range (min … max):    3.929 s …  4.022 s    10 runs
 
Benchmark 1: seq 4294967296 4295967296 | target/release/factor > /tmp/a
  Time (mean ± σ):      5.285 s ±  0.274 s    [User: 4.067 s, System: 1.231 s]
  Range (min … max):    5.103 s …  5.991 s    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Benchmark 1: seq 9223372036854775808 9223372036854875808 | target/release/factor > /tmp/a
  Time (mean ± σ):      2.703 s ±  0.011 s    [User: 2.547 s, System: 0.157 s]
  Range (min … max):    2.692 s …  2.732 s    10 runs
 
Benchmark 1: seq 18446744073709551616 18446744073709651616 | target/release/factor > /tmp/a
  Time (mean ± σ):      4.123 s ±  0.080 s    [User: 3.956 s, System: 0.162 s]
  Range (min … max):    4.053 s …  4.277 s    10 runs
 
Benchmark 1: seq 1267650600228229401496703205376 1267650600228229401496703206376 | target/release/factor > /tmp/a
  Time (mean ± σ):     11.681 s ±  0.129 s    [User: 11.623 s, System: 0.014 s]
  Range (min … max):   11.519 s … 11.857 s    10 runs
 
Benchmark 1: seq 170141183460469231731687303715884105728 170141183460469231731687303715884105828 | target/release/factor > /tmp/a
  Time (mean ± σ):     37.721 s ±  0.456 s    [User: 37.565 s, System: 0.020 s]
  Range (min … max):   37.072 s … 38.755 s    10 runs

Old version 

Benchmark 1: seq 2 1000000 | target/release/factor > /tmp/a
  Time (mean ± σ):      4.336 s ±  0.060 s    [User: 3.089 s, System: 1.260 s]
  Range (min … max):    4.276 s …  4.463 s    10 runs
 
Benchmark 1: seq 4294967296 4295967296 | target/release/factor > /tmp/a
  Time (mean ± σ):      9.089 s ±  0.157 s    [User: 7.692 s, System: 1.412 s]
  Range (min … max):    8.916 s …  9.376 s    10 runs
 
Benchmark 1: seq 9223372036854775808 9223372036854875808 | target/release/factor > /tmp/a
  Time (mean ± σ):     28.509 s ±  0.205 s    [User: 28.215 s, System: 0.213 s]
  Range (min … max):   28.273 s … 28.953 s    10 runs
 
Benchmark 1: seq 18446744073709551616 18446744073709651616 | target/release/factor > /tmp/a
  Time (mean ± σ):     34.087 s ±  0.326 s    [User: 33.756 s, System: 0.235 s]
  Range (min … max):   33.655 s … 34.523 s    10 runs

Benchmark 1: seq 1267650600228229401496703205376 1267650600228229401496703206376 | target/release/factor > /tmp/a
  Time (mean ± σ):     133.115 s ±  5.023 s    [User: 132.686 s, System: 0.054 s]
  Range (min … max):   121.333 s … 138.788 s    10 runs

Took too long but I got 1472 for a single run
1472.008

```